Add files via upload
HASHER-OOBE is a tool to download the devices HASH code using a flash drive and one simple command. ReadMe.txt and ReadMe.md have steps for beginners who may be lost and need a specific step-by-step tutorial on how to setup and complete a devices Autopilot enrollment on Microsoft Intune Admin Center.